chore: rename utility file to better reflect function
